sparse_coding/requirements_amd.txt
# AMD GPU / Windows-compatible requirements
# Removed: nvidia-* (CUDA-specific), triton (Linux-only), uvloop (Unix-only)
# torch, torchvision already installed via torch-directml setup — skip them here

aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.0
appdirs==1.4.4
async-timeout==4.0.2
attrs==23.1.0
# autobahn, Automat, Twisted, txaio, txtorcon, magic-wormhole, spake2, PyNaCl, hkdf, pyOpenSSL
# are Twisted/wormhole networking libs only needed for magic-wormhole file transfer — skip if unused
baukit @ git+https://github.com/davidbau/baukit@e14a18a6ad6cf9e0d6a5dc7a97e671e393c01682
black==23.3.0
blobfile==2.0.2
# boostedblob removed — it requires uvloop which is Unix-only
boto3==1.26.158
botocore==1.29.158
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==41.0.3
circuitsvis==1.41.0
cycler==0.11.0
datasets==2.12.0
dill==0.3.6
docker-pycreds==0.4.0
docstring-parser==0.15
einops==0.6.1
exceptiongroup==1.1.1
fancy-einsum==0.0.3
filelock>=3.12.0
fonttools==4.39.4
frozenlist==1.3.3
fsspec>=2023.5.0
gitdb==4.0.10
GitPython==3.1.32
graphviz==0.20.1
h11==0.14.0
hkdf==0.0.3
httpcore==0.17.2
httpx>=0.25.0
huggingface-hub==0.15.1
humanize==4.6.0
hyperlink==21.0.0
idna==3.4
importlib-resources==5.12.0
incremental==22.10.0
iniconfig==2.0.0
jaxtyping==0.2.19
Jinja2>=3.1.2
jmespath==1.0.1
joblib==1.2.0
kiwisolver==1.4.4
lxml==4.9.2
magic-wormhole==0.12.0
markdown-it-py==2.2.0
MarkupSafe>=2.1.2
matplotlib==3.7.1
mdurl==0.1.2
mpmath>=1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy==1.3.0
mypy-extensions==1.0.0
networkx>=3.1
# neuron-explainer: install manually from local modified automated-interpretability folder
numpy>=1.24.3
optree==0.9.1
orjson==3.9.1
packaging==23.1
pandas==2.0.2
pathspec==0.11.1
pathtools==0.1.2
Pillow>=9.5.0
platformdirs==3.5.3
plotly==5.14.1
pluggy==1.0.0
progressbar2==4.2.0
protobuf==4.23.2
psutil==5.9.5
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodomex==3.18.0
Pygments==2.15.1
PyNaCl==1.5.0
pyOpenSSL==23.2.0
pyparsing==3.0.9
pytest==7.3.2
python-dateutil==2.8.2
python-utils==3.7.0
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.31.0
responses==0.18.0
rich==13.4.1
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.2.2
scipy==1.10.1
sentry-sdk==1.24.0
service-identity==21.1.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
spake2==0.8
sympy>=1.12
# tables (PyTables/HDF5) — install separately if needed: pip install tables
tenacity==8.2.2
threadpoolctl==3.1.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
# torch, torchvision — already installed, do NOT reinstall (would pull CUDA version)
# torchopt==0.7.1 — may conflict with torch 2.4.1; install separately after verifying
torchtyping==0.1.4
tqdm==4.65.0
transformer-lens @ git+https://github.com/neelnanda-io/TransformerLens@ae32fa54ad40cb2c3f3a60f1837d0b4899c8daae
transformers==4.30.1
# triton — Linux/NVIDIA only, skip on Windows
# Twisted, txaio, txtorcon — only needed for magic-wormhole, skip if unused
typed-argument-parser==1.8.0
typeguard==3.0.2
types-requests==2.31.0.1
types-urllib3==1.26.25.13
typing-inspect==0.9.0
typing_extensions>=4.6.2
tzdata==2023.3
# uvloop — Unix-only event loop, skip on Windows
urllib3==1.26.16
wandb==0.15.3
xxhash==3.2.0
yarl==1.9.2
zipp==3.15.0
